Move Arbitrary ChainStateType into IsChainState
This is only temporary as the IsChainState type class might not be even
needed (IsTx is enough).
