+++ hydra-node/src/Hydra/Chain.hs
@@ -32,9 +32,9 @@ import Hydra.Tx (
   SnapshotVersion,
   UTxOType,
  )
+import Hydra.Tx.IsTx (ArbitraryIsTx)
 import Hydra.Tx.OnChainId (OnChainId)
 import Test.Cardano.Ledger.Core.Arbitrary ()
-import Test.Hydra.Tx.Gen (ArbitraryIsTx)
 import Test.QuickCheck.Instances.Semigroup ()
 import Test.QuickCheck.Instances.Time ()
 
@@ -177,7 +177,7 @@ deriving anyclass instance IsChainState tx => FromJSON (PostTxError tx)
 
 instance IsChainState tx => Exception (PostTxError tx)
 
-instance ArbitraryIsTx tx => Arbitrary (PostTxError tx) where
+instance (ArbitraryIsTx tx, IsChainState tx) => Arbitrary (PostTxError tx) where
   arbitrary = genericArbitrary
 
 -- | A non empty sequence of chain states that can be rolled back.
@@ -271,7 +271,7 @@ deriving stock instance (IsTx tx, IsChainState tx) => Show (ChainEvent tx)
 deriving anyclass instance (IsTx tx, IsChainState tx) => ToJSON (ChainEvent tx)
 deriving anyclass instance (IsTx tx, IsChainState tx) => FromJSON (ChainEvent tx)
 
-instance ArbitraryIsTx tx => Arbitrary (ChainEvent tx) where
+instance (ArbitraryIsTx tx, IsChainState tx) => Arbitrary (ChainEvent tx) where
   arbitrary = genericArbitrary
 
 -- | A callback indicating a 'ChainEvent tx' happened. Most importantly the
