+++ hydra-node/src/Hydra/HeadLogic/Input.hs
@@ -8,7 +8,7 @@ import Hydra.API.ClientInput (ClientInput)
 import Hydra.Chain (ChainEvent)
 import Hydra.Chain.ChainState (IsChainState)
 import Hydra.Network.Message (Message, NetworkEvent)
-import Test.Hydra.Tx.Gen (ArbitraryIsTx)
+import Hydra.Tx.IsTx (ArbitraryIsTx)
 
 type TTL = Natural
 
@@ -32,6 +32,6 @@ deriving stock instance IsChainState tx => Show (Input tx)
 deriving anyclass instance IsChainState tx => ToJSON (Input tx)
 deriving anyclass instance IsChainState tx => FromJSON (Input tx)
 
-instance ArbitraryIsTx tx => Arbitrary (Input tx) where
+instance (ArbitraryIsTx tx, IsChainState tx) => Arbitrary (Input tx) where
   arbitrary = genericArbitrary
   shrink = genericShrink
